Implementation of Session expiration.
This is probably a bit bigger than I usually like to have in a single
PR, but it got away from a bit in the end.

For this PR I've hardcoded the UDP session timeout to be 1 minute, but
eventually we should make this configurable.

So this PR does the following:
- Sessions track and `expiration` value, that is reset to now + timeout
  on each packet sent or received
- There is also pruning functionality that is run on every interval to
look for expired sessions, close their receiver loops and delete them
from the Session Map.