Implementation of Filter API and FilterRegistry
FilterRegistry is a way to register the default set of filters for a
client of server configuration (and then add your own if this is used as
a library).

This also includes an accompanying Filter trait implementation for UDP
traffic going in both direction in and out of the proxy.