Implementation of DebugFilter
This is an implementation of a Filter that pushes all input and output
to stdout as it flows through. Useful for debugging flow and data
operations. Also was a good use case to work through the Api surface of
the Filter trait.

This filter has already been added to the default Filter Registry as
well.

Note: This cannot yet be configured and put into a Client/Server Proxy.
This is the task I am tackling next.