Implementation of FilterChain
This is so we can take a set of Filters and string their execution in
the appropriate places.

FilterChain also implements Filter, so it's clear what FilterChain
functions match to Filter functions.

Once we have this in place, we can build a FilterChain
from the combination of a Config the FilterRegistry on startup of a
Quilkin executable.