Refactor Config.get_endpoints()
Not sure why I decided to have that function return a HashMap, but to
implement filters, we need access to the underlying Endpoints, since
they get passed into several Filter functions.

So refactored the function to return a Vec<Endpoint>, which will make
things much easier to work with.

Work on #1