Refactor UDP Test Utils to be more flexible
Refactored the UDP test utilities, such that they return the message
that is received, rather than expecting a standard "hello" message.

Since Filters will manipulate strings, we can't expect there to only be
one string value that gets passed through when testing. So let's return
it so we can test for values that are expected in the test.