Add client proxy load balancing support
This commit adds RoundRobin and Random load balancing
support to a proxy as well as config support for multiple
addresses and load balancer policy on the client proxy config.
The default behavior if no policy is set or is a server proxy
sends packets to all endpoints.

This also introduces the `rand` crate as a dependency, used by
the random load balancing implementation.

Resolves #3