Lazy instantiation of Filters
This PR implements lazy instantiation of filters through a closure -
which also leads to the ability to pass the server_yaml::Value for the
config of the filter, which was a nice side effect.

To prove the lazy eval and config passing was working, this also added
the ability to pass an "id" value to the DebugFilter, such that you
could implement several of them in a Filter configuration in between
other filters, and be able to identify which log statement is which
step.

This also led to finding multiple places in which an Arc was being used,
but only the guarantees of a Box where actually needed, so this was also
switched out.

An integration test for the DebugFilter is still incoming.

Work on #1