Convert connection_id's to base64 byte arraus
Connections id's should be Vec[u8] instead of strings, as the connection
authentication tokens could easily be arbitrary data rather than
strings.

This PR implements this in the config, as well as changes the yaml
storage to be a base64 encoding of the byte array.

Implementation is done through a new type `ByteArray` which implements a
custom de/serialiser for serde to create this functionality.

Deliberately didn't use serde's de/serialize_with attributes as I
couldn't find clean way to implement this in a way that would work for
both the Vec<u8> in the Client config as well as a Vec<Vec<u8>> in the
Server configuration.

Work on #8