Concat Byte Filter
The `ConcatBytes` filter's job is to add a byte packet to either the
beginning or end of each UDP packet that passes through. This is
commonly used to provide an auth token to each packet, so they can be
routed appropriately.

This also includes implementing `Filter` on `Box` to make reusable test
assertion functions easier to write.

Next step will be to remove `connection_id` from the Client config.

Work on #8