Upgrade to all dependencies.
This upgrades all dependencies, which includes some breaking changes to
Tokio, with adjustments to our code base to handle the changes.

For example: UdpSocket can no longer be split. The Tokio advice was to
wrap the UdpSocket in an Arc. Also, the API surface for Tokio watch
channels has also changed.