Implementation of Compression Filter.
Default implementation uses Snappy compression, but it has been
implemented such that we can add extra compression modes over time.

This includes adding ProxyMode to the `CreateFilterArgs`, as the
directions of compression/decompression change depending on the mode.

Next up are documentation and integration tests. After that we will
tackle an integration test (and corresponding fixes) wherein we run both
authentication filters and compression to validate correct ordering of
Filters on the `FilterChain`.