Fix rate limiter flaky test
Use polling where possible, otherwise reduce the interval
between refills (we get the same effect without increasing
the sleep time of the test)

`time::pause` doesn't affect scheduling unfortunately it'll only controls
the value of Instant::now which isn't used in this filter.
We can only poll in the cases where the test expects a transition from
no-tokens to some-tokens which isn't all of them.

Fixes #177