Align Filter trait with documented terminology
I was finding it hard to reconcile `on_upstream_receive` with the
(almost) documented `direction: downstream` and similar
`on_downstream_receive` with `direction: upstream`.

Therefore, I refactored `on_upstream_receive` to `on_downstream` and
`on_downstream_receive` to `on_upstream`, as well as renaming the
context and response objects.

I found this easier to understand as the Filter event is processing
data either moving upstream or downstream. I no longer have to think
about where it was received from -- the bytes are just moving in one
direction or other.

I feel like now the Filter pattern directly matches what we have in
our documentation, and there is no contextual mismatch, and it will be
much easier to learn and write filters for.