Release and Debug image release pipeline
This appends to the release `cloudbuild.yaml` to create a release and
debug tagged image when run based on the `distroless` image.

This also includes several steps that are needed for ongoing internal
compliance - such as maintaining a list of dependencies and their
licences, zipping MPL licenced code, etc.

Fixes #201