Add /live health endpoint to admin server
Create a `Health` module which tracks if a panic occurs
anywhere in the code base (which may or may not be on the main thread),
and moves the system to unhealthy.

In the future we could add extra checks to this module as we discover
more things that impact proxy health.

Closes #73