Merge sessions map and pruning logic
Currently sessions management logic is mixed in with the actual server
so this moves it out to a separate SessionManager.

I originally wanted to re-use this SessionManager to add
session affinity to the load balancer but quickly realised that this
wouldn't be possible currently because looking up a session is
unfortunately async (due to locking) and filters aren't async. Not
currently sure how we'd want to tackle this issue, atm at least would rather not
do an async Filter trait since that brings other issues and we might
end up with a perf regression spawning a new task per packet