Usage documentation
Includes documentation for execution of the release binaries and docker
images.

This also includes removing this from the README.md page, as well as the
build information (which is still in CONTRIBUTING.md), as the target for
the documentation is based around the release binaries for the upcoming
alpha release.