Quickstart for Agones + Quilkin
Utilises the Agones example to do a walkthrough of running Xonotic with
a sidecar, and also to do client side compression to server side
decompression.

Note: Won't work out of the box until we do our initial release. To test
edit the sidecar.yaml and sidecar-compress.yaml to point at a dev image
of Quilkin in your project.

Work on #62