Wait for execution result channel before recving
If we fail to fetch connect to the xds server on startup, there's
a race between the xds client placing the error on the channel
and the resource manager dropping the receiver due to no longer
being able to receive updates from the client.
If the latter happens first, we get a generic 'channel dropped' error
message instead of the actual error message.

This adds a short sleep before checking the channel for an error message
to give the client enough time to send the error.

Also return a result from main rather than unwrapping. Doesn't change
much beyond the error message being less scarier if it doesn't show
up as a panic.