Example Grafana dashboard for Quilkin Metrics
Created setup scripts (Makefile) and a sample dashboard (screenshot in
this commit) to show off some of Quilkin's expected metrics.

Over time this can be expanded to include more panels in this dashboard,
and other dashboards as well 👍🏻

At this stage, stayed away from Filter specific metrics, so that we have
something that works out of the box, and can be expanded as needed.

Closes #234