Make + Docker toolchain for development
Was too difficult to create a container image for a local development
version of Quilkin. Also if you wanted to cross compile, that required
knowledge of `cross`, and likely more in the future.

This provides a Make + Docker toolchain that is an enhancement for the
current local Rust development in that it can:

* Run all the tests tha CI runs
* Cross compile to other OS's for a local dev version.
* Create a Container image for a local dev version.

In this commit I deliberately haven't replaced any of the CI and release
toolchain, but that is coming next. This same build image will replace
the CI buildstep image, and the make targets that can be used here for
development will be used for both CI and release - so there is a
consistent experience across all layers of development.

Also next will be trying to get cross compilation to happen for macOS
from Linux, without having to install any extra tooling locally as well.

Work on #241
Closes #296