Tweaks to release checklist.
I realised that the checks for issues with `cargo publish` should occur
before the PR's to update version numbers, so that it can be fixed in
that pull request.

Also realised that publishing to crates.io should happen before
submitting the release on github, just in case something goes wrong, the
release is not stuck in a half-released state.