Bug: Changelog didn't have the version
I introduced a bug when refactoring the build pipleline, wherein there
wasn't a `./version` file anymore which stored the current version
number, such that it could be shared across build steps.

Therefore there was no version on the changelog! This fixes that by
creating a `version` taget for the Makefile and pushing that into a file
the changelog generator can use.

Work on #313