Documentation snapshot/release README.md section
Break out a section on the README.md linking to tagged snapshots of
documentation, so that as API changes occur it's still easy to see
previous API documentation and accompanying guides. It also stops new
users from getting confused by in-flight API changes.

I believe this will aso unblock #293 and also remove the requirement to
get a 0.1.0 snapshot for #319, which could block other PRs as well.