`make docs` to preview documentation locally
I had some updates to the documentation I wanted to make, but I wanted
to be able to preview the whole documentation set created by `cargo doc`
and `mdbook` in the the same directory structure as we have it hosted on
Github pages.

Also, I saw #363.

So with some extra components in our build image, and a combination of
`cargo watch` and a python dev library that hosts the files, and hot
reloads on changes for all documentation, we have a nice local tool for
writing and previewing documentation locally!

Closes #365