CI: Add cross compile and basic image test
Identifying that we accidentally keep breaking either the release
process and/or the released container images since we don't automate
testing of it.

So this PR does several things:

* Fix the issues I found in compiling for macos when testing this CI
improvement.
* Adds cross compilation into the CI process.
* Add building the image into the CI process.
* Run the Quilkin release and debug images against their default
configuration (which does nothing) for a few seconds, to make sure they
don't error.
* Increase build image to 32 cores, which keeps the CI build process to
~13 minutes, which is not too bad.

Long term, if we wanted to keep images/binaries temporarily for PR
builds, we can also do so.

Work on #318