Add support for versioned filter chains
As discussed in #401, this adds support to specify multiple filter
chains and have quilkin strip off a byte slice
from packets and compare that against several filter chains
to find a filter chain match that would be used to process the
packet.
The updated docs and tests should show how this works from a user's pov.
Config wise, it works similar to what was proposed in the ticket, the
only difference being that confuring the single filter chain we have
today lives 1 level nested at `static.filter_chain.filters` as opposed
to `static.filters` that we have today.

Resolves #401