Add a rate limited logger
We currently have a `LOG_SAMPLING_RATE` variable which is a more naive
solution to limit the messages we log. Instead, this
makes sure that we don't log over a threshold across the proxy in a
single time window.

Impl wise its a tiny wrapper (log.rs) around the log macros that check if we're
allowed to log a message and drops it otherwise. The remaining changes
update usages so even though this touches a few files its not as large a
change