Docs: Fix broken links (#445)
This fixes for all the broken links I found while working on #367.

Next PR to close #367 will include automated testing of the
documentation, just waiting to see if a bug fix I have in PR will flow
through in a timely manner. If not, will install in our build image from
my fork with the fixes.