CI: Documentation test gen, links, images, etc.
Introduces both building the documentation sites with `mdbook` and
`cargo doc` as well as placing them in the same directory structure as
they exist in on GitHub pages.

Utilising `htmltest` to test the links, images, scripts etc in the
`mdbook` pages (there are lots of invalid anchors in the rustdoc, so not
 testing the links for now - but we could later).

To use locally: `make test-docs` will run the tests for you.

Extra notes:
* Had to use a fork of `htmltest`, as awaiting the merging of a critical
  bugfix.
* Split up each of the test targets in `cloudbuild.yaml`. This requires
  a bit more work when writing new test targets in the Makefile, but
  will be easier to determine why things go wrong in CI.
* Broke up the Dockerfile into sections, to make it faster for local
  rebuilding.

Fix #367