archive_dependencies.sh: Handle no dependencies
If there are no dependencies utilised that we need to archive source for,
this breaks the build process since no dependencies-src.zip file is
created, and the Dockerfile has nothing to copy.

This changes the script such that it starts with an empty zip file, and
then appends dependencies if and when they exist.

Unblocks #457