Benchmark comparing read and write throughput
Wanted to be able to highlight if we had bottlenecks in performance on
read vs write operations on the proxy.

This adds an extra benchmark to throughput.rs called "readwrite" and
follows a similar pattern as the overall throughput benchmark, with both
direct and proxies traffic utilised as extra comparison values.

Work on #410