Add short poll to xDS clients as workaround, improve xDS tracing
Currently for some reason the management server is not sending back to
active clients changes because there is seemingly no receivers. This is
a temporary workaround to have the client constantly polling for new
changes so that will stay up to date, while we work on addressing the
management's bi-directional communication.