Total review of guide
Sorry that this is a heavy PR, but once I started hacking on this,
all the changes ended up running into each other, so it was hard to
break it all apart.

The focus was primarily on making sure it was easy to understand the
difference between "proxy mode" and "xds provider mode", fix bugs in the
docs, update based on new features and improve general onboarding, as
well as organise more of the documentation into folders.

This included:

* Add Agones to introduction
* Put quickstarts in their own folder
* Fix bug in ncat quickstart.
* Rearranging the TOC to be clearer on proxy mode vs xDS.
* Updates to Proxy Concepts
* Mode top level metrics pages for proxy and xDS.
* Update integration architectures to be prettier.
* Updated integrations with new features.
* Made headers semantically ordered across the docs.
* Updates to the admin documentation.
* Update FAQ
* Split up xDS documentation into seperate sections
* Generate and include command line help output.
* Full update on Agones xDS provider inline with code functionality.

Filed a few more issues on some extra documentation I would like to
create at a later date, but I think this gives us a strong foundation to
build on.