Consistency of "packets_dropped_total" metric
Per https://prometheus.io/docs/practices/naming/#metric-names

"A metric name should have a suffix describing the unit, in plural form
Note that an accumulating count has `total` as a suffix, in addition to
the unit if applicable."

Therefore, all "packets_dropped" metric names should be
"packet_dropped_total".

This implements that change across all packets_dropped metric (which as
far as I can tell is the only metric missing this suffix) for both the
metric and variable name, which was sometimes implemented and sometimes
not.