Timeout for xDS gRPPC Client connection
I can't seem to find any default timeout for connection from a tonic
gRPC connection - which leads me to think it will block forever.

This might be the cause of why there is only a single retry in #660, but
at the very least, we should have a timeout anyway, just in case a
connection ends up deadlocking for whatever reason.

Work on #660