Error for invalid endpoint
Rather than searching for a string in an error message, this creates an
Error type for when an xDS endpoint address is invalid.

`http::Uri` will allow some Uri formats that will fail for us, so we
also add some extra checks of our own to ensure we have all the
information we need to make a connection.