Make: Even better package_version
#675 was better than the original, but still potentially brittle if the
format of the Cargo.toml changed.

This is better than #675, since gets the value from a `cargo` command,
so it no longer matters if the Cargo.toml changes -- everything should
still work as always, since it's a format.

We are also now only relying on `cargo` and `awk` as dev dependencies,
which is actually one less dependency than before (grep).