Change quickstart example README.md to point to docs.
We had duplicate documentation in both our quickstarts and the in the
folders for their code references as well.

Changed the README to just specify which Guide title to look for in the
docs.

I didn't link them directly, as links change, and we don't know which
version of Quilkin the end user is using.