Fix dead lock in SessionPool
The problem was that in dashmap we were occasionally getting unlucky and
having sessions sometimes end up in the same shard, which would cause a
deadlock since one would be trying to insert into a shard while we still
hold a reference to one.

Fortunately this is a pretty simple fix of not returning a reference at
all and instead just returning the socket, since that was the only thing
that was used out of the session struct.