Remove built dependency
This was pulling in git2 and writing a lot of data that was never used
in the crate. This just changes build.rs to manually read the HEAD
commit information and set an env var.
AFAICT this is actually superior in every way to built since we no
longer have the extra dependencies, and we actually get up to date
information when the commit changes, which AFAICT was not the case with
built since it was not registering file changes.