Move game traffic sockets to io-uring (#850)
* Move game traffic sockets to io-uring

* Don't panic if io-uring thread fails to initialise

* Fixup rebase

* Keep entry count

This just changes the entry count to be kept as a single atomic, rather
than iterating and summing the entire map for every call to
num_of_endpoints/has_endpoints. Changes quilkin proxy loop from ~90MB/s
to ~125MB/s, so a fairly good increase for little work

The count may not be exactly accurate, but it's more of a boolean than
an exact number anyway

* Fix non-linux build

* Fixup tests

Several tests were using the same port and would sporadically fail on my
machine, this just changes it so that all of the tests (some already
did this) bind to an ephemeral port unless otherwise specified, then
waits until all of the workers have spawned before sending back which
port was assigned. This also adds a configuration option to specify the
number of workers to use as running all the tests in parallel where each
uses the max number of workers also caused tests to unreliably fail due
to contention

* Add simple buffer pool

This improved the benchmarks slightly, not as much as I had hoped, but
this initial implementation is super simple and can be improved.

One behavior change is that the snappy compression is no longer using
streaming compression in favor of just compressing a single chunk as
streaming compression is overkill for the size of data in a typical
packet, in fact one of the tests had to be changed to add more data to
the test packet as otherwise the packet could still be turned into a
utf8 string since the block compression only added a size header due to
how small it was

* Fix doctests

* Fix example

---------

Co-authored-by: Jake Shadle <jake.shadle@embark-studios.com>