Remove protoc dependency
This removes protobuf-src as a dependency and default feature in favor of committing the generated rust files instead. On my machine this took a clean release build from ~80s to ~60s. To me it feels like the builds are still far slower than they should be but this is a good easy win.