Improve compile times (#896)
* Remove unnecessary proc macros

* Update maxmind

* Remove protoc dependency

This removes protobuf-src as a dependency and default feature in favor of committing the generated rust files instead. On my machine this took a clean release build from ~80s to ~60s. To me it feels like the builds are still far slower than they should be but this is a good easy win.

* Update generated

* Replace chrono with time

* Rename ServiceClient::connect -> connect_to_endpoint

If protobufs are generated without transport (and thus no ::connect method) the previous naming would cause infinite recursion rather than a compile time failure

* Readd copied protobufs

Apparently git didn't like me removing submodules and changing the contents of the folders at the same time

* Remove commented out code

* Oops

* Validate protobufs in CI

* Format

* Oops

* Cancel on push

* Sigh, double timeout

* Fix tarball extraction

* Double again

* Sigh

* Sigh x2

* Install protoc because ubuntu is the worst

* Sigh

Guess cloudbuild doesn't set this

* Unzip shenanigans

* uhm what

* ...

* just bump everything

* Remove comments

* Update proto-gen version

* Remove v prefix

* Ugh

* ...

* Ignore flaky test for now

* Fix agones