Move the rest of the CLI into components
This moves the rest of the CLI commands into components the same as proxy in the previous commit.

The goal of this change is to start to tease apart the user interface (CLI) portions from the actual code that is run, to make writing tests easier

This also completely removes the notion of unwrapping the `RuntimeConfig` from the `Admin` inside various places in favor of passing in the actual config for the component in a type safe manner. For instance previously running the manage command with a relay server set would panic as .unwrap_agent() was used inside MdsClient