Move xDS and protobuf definitions into separate crates.
This commit moves the xDS functionality in its own library, with the
eventual goal of making the xDS component its own library that can be
used by other services, currently this just extracts the functionality
as-is, with the only change being the addition of the
`xds::config::Configuration` trait which abstracts over the usage of
`config::Config` in xDS as I wanted to keep that in the Quilkin library.

The `quilkin-proto` library was also added to allow for the protobuf
definitions to be shared between the xDS library and the quilkin
application.