Update crates
Unfortunately this isn't as big of a win as I would have hoped, primarily due to tonic and hickory-resolver both using many outdated crates, but it is a step in the right direction.

hyper -> 1.3 - This aligns better with some crates eg. kube-client, but misaligns us with tonic/axum. This is also annoying because 1.0 re/moved a lot of functionality so we have new dependencies on small utility crates, but it is what it is
trust-dns-resolver -> hickory-resolver - trust-dns-resolver was a major source of outdated crates, hickory-resolver, being a fork of that project, carries on the tradition, though is _slightly_ better, as well as just being the future as opposed to trust-dns