Remove dynamic error properties
This removes the dynamic/heap allocated properties from PipelineError and all of its subtypes. This means we lose some information, notably in the case of TokenRouter::Error no longer having the token info, but means we remove the chance of unbounded (well, bounded by the size of the token as a base64 string) heap growth in the presence of completely random packet data that can't be routed to an endpoint.