First pass implementation of XDP I/O loop (#1066)
This PR is the real meat of the XDP implementation, the I/O loop.

The setup of the memory, sockets, and rings is done in the root xdp module, along with the spawning of one (named) thread per NIC queue, the actual packet processing code is in the xdp/process module. This split of code allows much easier testing, since we can just have a Umem (memory map) and don't need to do any actual I/O.

For the sake of simplicity in this initial pass I decided not to share any code at all with the current sessionpool, instead opting for a more minimal approach to pairing client <-> server sessions by having a single TTL map for all servers, where each unique client gets its own port in the range of valid ephemeral ports that Linux itself doesn't assign, which gives 4535 unique clients per server, so the TTL applies to the server instead of each individual client. Other than that difference, the only thing different about XDP versus the other implementations is that packets need to parsed manually when received to pull out the source information, and, unless being routed to multiple servers, the packet is then modified inline and queued for sending with the new source and destination information as determined by the filters.